+++ lumber-deploy.js
@@ -5,4 +5,4 @@ program
   .description('Deploy your back office to your production environment.')
   .parse(process.argv);
 
-console.log(`Read the instructions: \n ${chalk.green('https://doc.forestadmin.com/developer-guide/lumber.html#deploy-to-production')}`);
+console.log(`Read the instructions: \n ${chalk.green('https://docs.forestadmin.com/lumber/getting-started/setup-guide#deploy-to-production')}`);